refactor(api): unificar endpoints de canales rss en un solo archivo
