+++ newsradar_api/app/api/routes/rss_channels.py
@@ -1,10 +1,20 @@
-from typing import List
-from fastapi import APIRouter, Depends, Response
+"""Este módulo define los endpoints relacionados con la gestión de canales RSS."""
 
-from app.schemas.rss import RSSChannel, RSSChannelCreate, RSSChannelUpdate
+from typing import List
+from fastapi import APIRouter, Depends, Response, HTTPException, status
+from sqlalchemy.orm import Session
+
+from app.schemas.rss import (
+    RSSChannel,
+    RSSChannelCreate,
+    RSSChannelUpdate,
+    RSSChannelResponse,
+)
 from app.schemas.user import UserInDB
 from app.stores.memory import rss_channels_store
-
+from app.database.database import get_db
+from app.services.rss_service import create_rss_channel, get_all_rss_channels
+from app.api.dependencies import get_current_gestor, get_current_user
 from app.utils.deps import get_current_user
 from app.utils.rss_utils import (
     ensure_information_source_exists,
@@ -16,6 +26,42 @@ from app.utils.rss_utils import (
 API_PREFIX = "/api/v1"
 
 router = APIRouter()
+
+
+@router.post(
+    "/",
+    response_model=RSSChannelResponse,
+    status_code=status.HTTP_201_CREATED,
+    dependencies=[Depends(get_current_gestor)],
+)
+def crear_canal_rss(rss_in: RSSChannelCreate, db: Session = Depends(get_db)):
+    """
+    Crea un nuevo canal RSS en el sistema.
+    [SOLO GESTORES] - Bloqueado a Lector usando la dependencia get_current_gestor.
+    """
+    try:
+        nuevo_canal = create_rss_channel(db, rss_in)
+        return nuevo_canal
+    except Exception as e:
+        # Aquí podrías capturar IntegrateError si la URL ya existe en BD, por ejemplo
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f"No se pudo crear el canal RSS: {str(e)}",
+        ) from e
+
+
+@router.get(
+    "/",
+    response_model=List[RSSChannelResponse],
+    dependencies=[Depends(get_current_user)],
+)
+def listar_canales_rss(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
+    """
+    Obtiene todos los canales RSS registrados.
+    Este endpoint sí es accesible por Lectores, el control de acceso es solo de creación.
+    """
+    canales = get_all_rss_channels(db, skip=skip, limit=limit)
+    return canales
 
 
 @router.get(
